refactor(pos): flatten service selection into a unified real-time cart
Service selection required opening a modal, adjusting quantity in a
separate staging cart, then clicking "Tambah ke Pesanan" to merge it into
the real cart FloatingOrderSummary shows - two pieces of cart state for the
same items, plus a whole extra open/close-modal round-trip for the most
frequent action in the flow.

Rename EnhancedServiceSelectionPopup to InlineServiceSelector and drop its
Dialog wrapper and staging cart: services/products now write straight into
the real cart on click (onAddService), landing at quantity 1 or 1kg with
further adjustment via the exact-value input already in
FloatingOrderSummary's cart list. Custom items keep their per-item drafting
fields (still need multiple fields filled before they're valid) but each
draft gets its own "Tambah ke Pesanan" action instead of a global batch
confirm.

In EnhancedLaundryPOS, handleServicesAndItemsSelected is replaced by
addServiceToOrder/addCustomItemToOrder writing directly to currentOrder/
dynamicItems, isServicePopupOpen is gone (the section has no open/closed
state now), and FloatingOrderSummary's "+" button scrolls the always-visible
section into view instead of opening a dialog.

CashPaymentDialog, OrderSuccessDialog, and ThermalPrintDialog are
unchanged - only the pre-payment item-selection flow was flattened.
